+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - M68k Browser: new tab "Directives" with reference documentation
+- M68k Browser "Amiga Hardware" tab: add links between registers in reference documentation
 
 ### Fixed